allow specifying alternate SVG converter
rsvg-convert is broken on Arch Linux for unknown reasons and has
remained so since before November 2015 (issue #180).

The prior commit (64397c2e) changes the resizing operation rsvg-convert
uses to zooming (works identically to specifying dimensions as tested
with rsvg-convert in Jessie), while this one allows specifying an
arbitrary alternate converter, as long as it is compatible with
rsvg-convert's syntax.

As a final piece to the solution, I have altered the nuvola-app-*-git
PKGBUILDs for Arch to specify lasem-render, which is included in [extra]
on Arch and has been tested to produce properly working icons.

[1] https://wiki.gnome.org/Projects/Lasem